% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_pool.R
\name{db_pool_disconnect}
\alias{db_pool_disconnect}
\title{Close the DuckDB pool}
\usage{
db_pool_disconnect(pool)
}
\arguments{
\item{pool}{A pool object returned by \code{\link[=db_pool_connect]{db_pool_connect()}}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Close the DuckDB pool
}
