% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pivot.R
\name{.col_key}
\alias{.col_key}
\title{Build a column key string from entity + period_label}
\usage{
.col_key(entity, period_label)
}
\description{
Build a column key string from entity + period_label
}
\keyword{internal}
