% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_auth.R
\name{is_admin}
\alias{is_admin}
\title{Check whether a role string is admin}
\usage{
is_admin(role)
}
\arguments{
\item{role}{Character role string from \code{\link[=get_user_role]{get_user_role()}}.}
}
\value{
\code{TRUE} / \code{FALSE}.
}
\description{
Check whether a role string is admin
}
