% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_browse.R
\name{mod_browse_server}
\alias{mod_browse_server}
\title{Catalog browser server}
\usage{
mod_browse_server(id, pool, user_role = "standard")
}
\arguments{
\item{id}{Module namespace id.}

\item{pool}{Pool object from \code{\link[=db_pool_connect]{db_pool_connect()}}.}

\item{user_role}{Reactive or plain character with the current user role.}
}
\description{
Accessible to both standard users and admins (FR-BR-00).
}
