% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the fin.extract Shiny application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}
}
\description{
Run the fin.extract Shiny application
}
